Fix workout exercises delete query

+++ app/Http/Controllers/ExerciseWorkout.php
@@ -23,18 +23,19 @@ class ExerciseWorkout extends Controller
             'user_id' => Auth::user()->id,
         ]);
 
-        return redirect()->back()->with('status', 'Exercise has been added to your workout!', 'test');
+        return redirect()->back()->with('status', 'Exercise has been added to your workout!');
     }
 
     public function destroy(Request $request)
     {
         // Check if workout is from the specific user, if it is, let them delete the exercise from it
         // Otherwise, deny the request.
-        if(workouts::where('id', $request->workout_id)->where('user_id', Auth::user()->id))
+        if(exercise_workouts::where('workout_id', $request->workout_id)->where('user_id', Auth::user()->id)->where('exercise_id', $request->exercise_id)->get()->isNotEmpty())
         {
-            exercise_workouts::where('exercise_id', $request->exercise_id)->delete();
+            exercise_workouts::where('workout_Stid', $request->workout_id)->where('user_id', Auth::user()->id)->where('exercise_id', $request->exercise_id)->delete();
             return redirect()->back()->with('status', 'Exercise has been deleted from your workout!');
         }
+
         return redirect()->back()->with('status', 'Nice try ;)');
     }
